Refactor imports in multiple components to use helper functions for price and date formatting
